+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 🚀 New features
 
-- Import PKHeX CSVs, custom CSVs or PkHeX tables to automatically populate the tracker. Click on Import in the sidebar to see how it works!
+- Import PkHeX tables to automatically populate the tracker. Click on Import in the sidebar to see how it works!
 
 🐛 Bug Fix